Initial commit
The main file in this project is the Join service, which implements mergeJoin
and hashJoin; everything else is just there to demonstrate/test this, and will
probably be moved or replaced/augmented with unit/e2e tests instead.

mergeJoin is currently limited to requiring that no duplicate rows (insofar as
rows that are equal with respect to the comparator passed into mergeJoin)
exist; otherwise, too few rows are returned. The next version will have this
fixed, and hashJoin doesn't suffer from this.