Empty arrays: optimized in joins, fixed bug in groupBys
In hashJoin and mergeJoin, checks are made for empty arrays at the start and
if either input is empty, the other array has callback applied as appropriate
and is returned. This was an optimization, as empty arrays would have been
returned anyway, but after unnecessary calls to hashFcn & comparator,
respectively.

In sortGroupBy and hashGroupBy, a check on the input for an empty array at the
start and if true, an empty array is immediately returned. This is a fix, at
lease for sortGroupBy, as previously, an attempt was made to access a[0], which
resulted in an out of bounds error if the input was empty.