Added Fluent API demo page; updated Formatter service
The Usage page was rewritten to describe the new query object and the fluent
API that goes with it in general. A Fluent API demo page was also added, which
shows a single complex query, along with the intermediate results, using the
.inspect() function.

The Formatter service was updated to include a toFilteredSource function, which
lets the user extract specific portions of a function's source. This was needed
to hide the calls to .inspect() --for brevity's sake-- in the Fluent API demo.