Made static versions of all fluent functions (except visit & execute)
The static functions are on the Join service and are called identically to
their non-static versions, except the 1st argument is the starting array and
the return value is the final array.

This was implemented by creating a prototypical query object, and for each of
its functions (excluding 'visit' & 'execute'), wrapping it in a new function
that calls the fluent version (after extracting the 1st argument to pass into
selectFrom), calls execute() on the returned query, and returns that.