Accepted strings/arrays as comparator/hashFcn/select callbacks
Fixes #4

Functions that accept comparators and hashFcn, and the callbacks of select and
selectFrom, are now checked for whether they are strings or arrays; in either
case, a new function is returned that implements logic like the following:

Comparator functions:

  query.orderBy('x')
  query.orderBy(['x', 'y'])

which respectively equal:

  query.orderBy(function(e1, e2) { return e1.x - e2.x; })
  query.orderBy(function(e1, e2) {
    return (e1.x - e2.x) === 0 ? (e1.x - e2.x) : (e1.y - e2.y);
  })

Hash functions:

  query.hashJoin(right, 'x', callback)
  query.hashJoin(right, ['x', 'y'], callback)
  query.hashGroupBy('x', callback)
  query.hashGroupBy(['x', 'y'], callback)

which respectively equal:

  query.hashJoin(right, function(e) { return e.x; }, callback)
  query.hashJoin(right, function(e) {
    return JSON.stringify([e.x, e.y]);
  }, callback)
  query.hashGroupBy(function(e) { return e.x; })
  query.hashGroupBy(function(e) { return JSON.stringify([e.x, e.y]); })

Select functions:

  Join.selectFrom(input, 'x')
  Join.selectFrom(input, ['x', 'y'])
  query.select('x')
  query.select(['x', 'y'])

which respectively equal:

  Join.selectFrom(input, function(e) { return {x: e.x}; })
  Join.selectFrom(input, function(e) { return {x: e.x, y: e.y}; })
  query.select(function(e) { return {x: e.x}; })
  query.select(function(e) { return {x: e.x, y: e.y}; })

An issue with the comparators is that reverse ordering cannot be specified.
A future commit might accept a minus sign as the first character to mean
reverse ordering (e.g. '-x' <==> `function(e1, e2) { return e2.x - e1.x; }`).