Added localeCompare option for string comparison
The options object for mergeJoin, sortGroupBy, and orderBy/sort now supports a
localeCompare boolean property (default false). If true, it uses the
localeCompare to do the comparison; otherwise, it does a Unicode code point
comparison (i.e. the < and > operators). The localeCompare options is much
slower, but necessary in cases where locale ordering is important.

The testspec for Join was updated to include string-keyed joins, with and
without the new localeCompare option.