Refactored JoinQuery constructor; added missing options param
JoinQuery was refactored to that its constructor takes all parameters needed to
define it, rather than just the previous query/array. This simplifies the fluent
function definitions, which now just call the JoinQuery constructor and return
its value. Also, instead of an array of [opFunction, opParams...] defining
the query's operation, the requisite parts are now their own members of the
JoinQuery object (namely, this.op and this.params), making the call to
execute() less complex.

Also, the fluent mergeJoin and sortGroupBy methods were not using the options
parameter passed into them. This was fixed.