Added caching & force option; updated README
JoinQuery now caches the results of the last execute call by default and
returns them rather than re-executing the query. To disable this for a
particular execution, the boolean `force` flag was added to execute's options,
which disables use of the cache (but not the updating of the cache) if set to
true.

The README was updated with the above, as well as removing mention that
the JoinQuery object is immutable (it's strictly not, due to the cache).
Any mention of the async-mode's promise having notify events was also removed,
as this has been deprecated (while it could be done, it might cause confusion
when caching causes some steps to be performed, but not others. The same problem
exists with the inspect() operation, which has been documented for now.