Refactored to accept JoinQuery as parameters
Technically, any parameter can now be a JoinQuery and it will get executed
& its result used where the parameter is expected. But, it's probably confusing
to do this for anything other than the right arrays for hashJoin and mergeJoin,
so the README has only been updated for those functions.

One issue that doesn't need resolving right now is that in sync mode, the input
is checked/executed before any parameters, while in async mode, parameters are
checked first. This can be simplified to enforce a certain order, but cases
where this matters at all should be rare, so this isn't a priority right now.